removed the old checkout
